requirements-dev.txt
pathspec==0.12.1
packaging==24.0
black==24.4.2
mypy-extensions==1.0.0
click==8.1.7
pyflakes==3.2.0
flake8==7.0.0
pycodestyle==2.11.1
pytest-django==4.8.0
iniconfig==2.0.0
pytest==8.2.1
pluggy==1.5.0
factory-boy==3.3.0
six==1.16.0
python-dateutil==2.9.0.post0
Faker==25.2.0
parso==0.8.4
mccabe==0.7.0
platformdirs==4.2.2
isort==5.13.2
jedi==0.19.1
docstring-to-markdown==0.15
importlib_metadata==7.1.0
ujson==5.10.0
whatthepatch==1.0.5
pytoolconfig==1.3.1
python-lsp-server==1.11.0
zipp==3.19.2
python-lsp-jsonrpc==1.1.2
dill==0.3.8
yapf==0.40.2
tomli==2.0.1
astroid==3.2.2
pylint==3.2.3
rope==1.13.0
autopep8==2.0.4
snowballstemmer==2.2.0
tomlkit==0.12.5
pydocstyle==6.3.0
